moved the auth service to the lib
